Clean up and add documentation to `MessagesPaginationStateHandling`
