+++ Sources/StreamChat/Utils/PaginationStateHandling/MessagesPaginationStateHandling.swift
@@ -4,49 +4,24 @@
 
 import Foundation
 
-struct MessagesPaginationState {
-    var oldestFetchedMessage: MessagePayload?
-    var newestFetchedMessage: MessagePayload?
-
-    var oldestMessageAt: Date?
-    var newestMessageAt: Date?
-
-    var isJumpingToMessage: Bool
-
-    var hasLoadedAllPreviousMessages: Bool
-
-    var isLoadingNextMessages: Bool
-    var isLoadingPreviousMessages: Bool
-    var isLoadingMiddleMessages: Bool
-
-    var isLoadingMessages: Bool {
-        isLoadingNextMessages || isLoadingPreviousMessages || isLoadingMiddleMessages
-    }
-
-    static var initial: Self = .init(
-        oldestFetchedMessage: nil,
-        newestFetchedMessage: nil,
-        oldestMessageAt: nil,
-        newestMessageAt: nil,
-        isJumpingToMessage: false,
-        hasLoadedAllPreviousMessages: false,
-        isLoadingNextMessages: false,
-        isLoadingPreviousMessages: false,
-        isLoadingMiddleMessages: false
-    )
-}
-
+/// A component responsible for handling the messages pagination state.
 protocol MessagesPaginationStateHandling {
+    /// The current state of the messages pagination.
     var state: MessagesPaginationState { get }
 
+    /// This function should be called before performing the pagination call.
     func start(pagination: MessagesPagination)
+    /// This function should be called with the result of the pagination call.
     func end(pagination: MessagesPagination, with result: Result<[MessagePayload], Error>)
 }
 
-class PaginationStateHandler: MessagesPaginationStateHandling {
+/// A component responsible for handling the messages pagination state.
+class MessagesPaginationStateHandler: MessagesPaginationStateHandling {
     var state: MessagesPaginationState = .initial
 
     func start(pagination: MessagesPagination) {
+        // When loading a page around a message we set isJumpingToMessage to true
+        // And only when we finish loading all the next messages, we set it back to false
         if pagination.parameter?.isJumpingToMessage == true {
             state.isJumpingToMessage = true
         }
@@ -83,92 +58,71 @@ class PaginationStateHandler: MessagesPaginationStateHandling {
 
         let oldestFetchedMessage = messages.first
         let newestFetchedMessage = messages.last
-        let oldestMessageAt = oldestFetchedMessage?.createdAt
-        let newestMessageAt = newestFetchedMessage?.createdAt
-
+        
         switch pagination.parameter {
         case .lessThan, .lessThanOrEqual:
-            state.oldestMessageAt = oldestMessageAt
             state.oldestFetchedMessage = oldestFetchedMessage
             if messages.count < pagination.pageSize {
                 state.hasLoadedAllPreviousMessages = true
             }
 
         case .greaterThan, .greaterThanOrEqual:
-            state.newestMessageAt = newestMessageAt
             state.newestFetchedMessage = newestFetchedMessage
             if messages.count < pagination.pageSize {
-                state.newestMessageAt = nil
+                state.newestFetchedMessage = nil
                 state.isJumpingToMessage = false
             }
 
         case .around:
-            state.oldestMessageAt = oldestMessageAt
-            state.newestMessageAt = newestMessageAt
             state.oldestFetchedMessage = oldestFetchedMessage
             state.newestFetchedMessage = newestFetchedMessage
+            if messages.count < pagination.pageSize {
+                state.isJumpingToMessage = false
+                state.hasLoadedAllPreviousMessages = true
+            }
 
         case .none:
-            state.oldestMessageAt = oldestMessageAt
             state.oldestFetchedMessage = oldestFetchedMessage
             state.newestFetchedMessage = nil
-            state.newestMessageAt = nil
-
             if messages.count < pagination.pageSize {
                 state.isJumpingToMessage = false
                 state.hasLoadedAllPreviousMessages = true
             }
         }
 
-        if let aroundMessageId = pagination.parameter?.aroundMessageId, !messages.isEmpty {
-            let midIndex = messages.count / 2
-            let midPoint: Int = Int(floor(Double(midIndex)))
-            let secondHalf = messages[midPoint...].dropFirst()
-
-            // If the message is in the mid point, it means there's more pages to load in both sides
-            if messages[midPoint].id == aroundMessageId {
-                state.isJumpingToMessage = true
-                // If the message is in the second half of the response, it means there are no more next messages,
-                // so it means jumping is finished.
-            } else if secondHalf.contains(where: { $0.id == aroundMessageId }) {
-                state.isJumpingToMessage = false
-                // Otherwise, if the message is on the first half, it means it loaded all previous messages.
-                // If the messages was not found, it could also mean we are jumping to a parent message.
-            } else {
-                state.hasLoadedAllPreviousMessages = true
-            }
-
-            if messages.count < pagination.pageSize {
-                state.isJumpingToMessage = false
-                state.hasLoadedAllPreviousMessages = true
-            }
-        }
-    }
-}
-
-class PaginationStateHandlerThreadDecorator: MessagesPaginationStateHandling {
-    let queue = DispatchQueue(label: "io.getstream.messages-pagination-state-handler")
-    let decoratee: MessagesPaginationStateHandling
-
-    init(decoratee: MessagesPaginationStateHandling) {
-        self.decoratee = decoratee
-    }
-
-    var state: MessagesPaginationState {
-        queue.sync {
-            self.decoratee.state
+        if let aroundMessageId = pagination.parameter?.aroundMessageId {
+            calculateHasLoadedAllMessagesBasedOnTheLocation(of: aroundMessageId, given: messages)
         }
     }
 
-    func start(pagination: MessagesPagination) {
-        queue.sync {
-            self.decoratee.start(pagination: pagination)
+    /// If we are jumping to a message we can determine if we loaded the oldest page
+    /// or the newest page, depending on where the aroundMessageId is located.
+    /// - If the aroundMessageId is in the middle of the messages response,
+    ///   it means there are still older and newer pages to fetch.
+    /// - If the aroundMessageId is on first half of the messages response,
+    ///   it means we loaded all the oldest pages
+    /// - If the aroundMessageId is on the second half of the messages response,
+    ///   it means we loaded all the newest pages
+    private func calculateHasLoadedAllMessagesBasedOnTheLocation(of aroundMessageId: MessageId, given messages: [MessagePayload]) {
+        guard !messages.isEmpty else {
+            return
         }
-    }
 
-    func end(pagination: MessagesPagination, with result: Result<[MessagePayload], Error>) {
-        queue.sync {
-            self.decoratee.end(pagination: pagination, with: result)
+        let midIndex = messages.count / 2
+        let midPoint: Int = Int(floor(Double(midIndex)))
+        let secondHalf = messages[midPoint...].dropFirst()
+
+        // If the message is in the mid point, it means there's more pages to load in both sides
+        if messages[midPoint].id == aroundMessageId {
+            state.isJumpingToMessage = true
+            // If the message is in the second half of the response, it means there are no more next messages,
+            // so it means jumping is finished.
+        } else if secondHalf.contains(where: { $0.id == aroundMessageId }) {
+            state.isJumpingToMessage = false
+            // Otherwise, if the message is on the first half, it means it loaded all previous messages.
+            // If the messages was not found, it could also mean we are jumping to a parent message.
+        } else {
+            state.hasLoadedAllPreviousMessages = true
         }
     }
 }
